:mute: Switch from println to debug logging.
Instead of printing out the "Rendering xx" line for progress using println! it now uses the debug! macro from the log crate which works better for libraries as it allows the user of the library to control which log output to get and where to put it (log file, stdout, stderr) and filter it and such.